+++ src/Exceptions/UnauthorizedException.php
@@ -1,0 +1,17 @@
+<?php declare(strict_types=1);
+
+/*
+ * This file is part of the Seriquynh package.
+ * (c) Quynh Xuan Nguyen <seriquynh@gmail.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sepiphy\Laravel\Acl\Exceptions;
+
+use Symfony\Component\HttpKernel\Exception\HttpException;
+
+class UnauthorizedException extends HttpException
+{
+    //
+}