+++ tests/HasRolesPermissionsTest.php
@@ -1,10 +1,8 @@
 <?php declare(strict_types=1);
 
 /*
- * This file is part of the Sepiphy package.
- *
+ * This file is part of the Seriquynh package.
  * (c) Quynh Xuan Nguyen <seriquynh@gmail.com>
- *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */